successfully integrated the updated TransactionController.java from Ezra's branch into main

+++ src/main/java/com/banking/service/UserService.java
@@ -41,7 +41,7 @@ public class UserService
     public Long getIdByUsername(String username)
     {
         User user = userRepository.findByUsername(username);
-        return Math.toIntExact(user != null ? user.getId() : -1);
+        return (long) Math.toIntExact(user != null ? user.getId() : -1);
     }
 
     public boolean exists(String username)
